Match NativeAOT parameter default handling
