+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System/Reflection/Runtime/General/NativeFormat/DefaultValueParser.cs
@@ -1,8 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Collections.Generic;
-
 using Internal.Metadata.NativeFormat;
 
 namespace System.Reflection.Runtime.General.NativeFormat
@@ -18,15 +16,6 @@ namespace System.Reflection.Runtime.General.NativeFormat
                     defaultValue = Enum.ToObject(declaredType, defaultValue);
                 return true;
             }
-
-            defaultValue = null;
-            return false;
-        }
-
-        public static bool GetDefaultValueFromAttributeIfAny(IEnumerable<CustomAttributeData> customAttributes, bool raw, out object? defaultValue)
-        {
-            if (Helpers.GetCustomAttributeDefaultValueIfAny(customAttributes, raw, out defaultValue))
-                return true;
 
             defaultValue = null;
             return false;
